+++ src/Ozymandias.cfg
@@ -91,7 +91,7 @@ logic:
     galvo:
         module.Class: 'galvo_logic.GalvoLogic'
         connect:
-            daq: 'DAQ'
+            galvoInterfuse: 'galvo_interfuse'
 
     ltune_laserlogic:
         module.Class: 'ltune_laser_logic.LtuneLaserLogic'
@@ -155,6 +155,11 @@ logic:
             flipper2: 'flipper2'
 
 hardware:
+    galvo_interfuse:
+        module.Class: "interfuse.galvo_interfuse.GalvoInterfuse"
+        connect:
+            daq: 'DAQ'
+
     purdue_laser:
         module.Class: 'laser.ltune_laser.LtuneLaser'
         options:
